requirements.txt
asgiref==3.8.1
Django
django-filter==24.2
djangorestframework==3.15.2
drf-yasg==1.21.7
greenlet==3.0.3
inflection==0.5.1
packaging==24.1
pandas
pytz==2024.1
PyYAML==6.0.1
SQLAlchemy==2.0.31
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1